pyproject.toml
[project]
name = "karma"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.7.0",
    "boto3>=1.35.0",
    "click>=8.0.0",
    "datasets>=3.6.0",
    "duckdb>=0.9.0",
    "evaluate>=0.4.0",
    "indic-transliteration>=2.0.0",
    "ipython>=9.3.0",
    "mysql-connector-python>=9.3.0",
    "pillow>=11.1.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.0",
    "rich>=13.0.0",
    "ruff>=0.12.0",
    "scikit-learn>=1.3.0",
    "torch>=2.7.1",
    "tqdm>=4.64.0",
    "transformers>=4.52.0",
    "weave>=0.50.0",
]

[dependency-groups]
dev = [
    "mteb>=1.38.32",
]

# Install the current package in editable mode
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
karma = "karma.cli.main:main"
